+++ CHANGELOG.rst
@@ -4,7 +4,8 @@
 - ElasticSearchUtility: removed obsolete methods get_by_uuid(), get_by_uuids(),
   get_by_type() and get_by_type()
   [masipcat]
-. ElasticSearchUtility: removed unused internal method _get_type_query()
+- ElasticSearchUtility: removed unused internal method _get_type_query()
+  [masipcat]
 
 
 7.0.0a5 (2021-07-30)